+++ docs/changes.md
@@ -17,6 +17,8 @@ hide:
 - New reset button for the *Custom colour values* in the Options.
 - Fixed error _Cannot read properties of null_ for new installs.
 - Monitor downloads handles different server replies better and provides a restart recommendation when toggled on.
+- **Text render** effect option is changed to use a toggle switch.
+- Changed the **Line Height** slider to display variations in percentage values.
 
 ## 5.3